+++ src/cli.ts
@@ -31,6 +31,7 @@ program
     .option("-l, --list", "list available data versions")
     .option("-w, --which", "show current data version")
     .option("-t, --history", "show migration history")
+    .option("-f, --force", "force migration")
     ;
 
 
@@ -102,7 +103,7 @@ export async function main() {
         }
     }
 
-    if (versions.indexOf(targetVersion) === -1) {
+    if (!program.force && versions.indexOf(targetVersion) === -1) {
         return fail("target data version not available");
     }