+++ CHANGELOG.md
@@ -6,6 +6,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+
+
+## 0.2.0 - 2018-08-21
+
 Added:
 
 - CLI options: